+++ src/spec/pages/Login.spec.tsx
@@ -1,7 +1,13 @@
 import React from 'react';
 import { MemoryRouter } from 'react-router';
 import renderer from 'react-test-renderer';
+import { act } from 'react-dom/test-utils';
+import { mount } from 'enzyme';
+import { AuthenticationDetails, CognitoUserPool } from 'amazon-cognito-identity-js';
 import Login from '@pages/Login';
+import Form from '@components/Form';
+import Input from '@components/Input';
+import Spinner from '@components/Spinner';
 
 describe('Login', () => {
 
@@ -19,6 +25,20 @@ describe('Login', () => {
     it('should render correctly', () => {
       testRenderer.toJSON();
       expect(testRenderer).toMatchSnapshot();
+    });
+  });
+
+  describe('handleAuthFlow', () => {
+    it('should display the spinner component after the submit button is clicked', () => {
+      const wrapper = mount(
+        <MemoryRouter>
+          <Login />
+        </MemoryRouter>
+      );
+      act(() => wrapper.find(Form).props().onSubmit());
+      wrapper.update();
+      const children = wrapper.find(Spinner);
+      expect(children.length).toEqual(1);
     });
   });
 });